fix: Add test for unique registration number
Add a simple test that a vehicle's RegistrationNumber is unique before attempting to add to the database.

I left a TODO comment to add the error message later after we have a system for such messages. For now it just redirects back to the page with the VehicleCreateViewModel filled in.