Minor changes in many files. The changes are: Correcting links Making sure the pages have only one title and one that makes sense (eg. "Register Vehicle" instead of "Create") Turning links into buttons, primary for saving and if there is only one, secondary for going back to list.
There is one thing NOT fixed in this PR:
There are two lists of vehicles, Index and ParkedVehiclesOverview.  This is a bit confusing. We are only linking to ParkedVehiclesOverview (not sure if those are only parked cars or all cars??) but with every action on vehicles (create, edit etc) the return view is Index. There is not enough time to change this, as making changes in controllers requires a lot more testing than just the razor pages. Also, there are some links on the Index page that does not exist on the ParkedVehiclesOverview (and possibly more cars, if there is a filter on parked cars only).